Updated DeleteEmployeeService

+++ ShiftLogger.Console/Presentation/Services/DeleteEmployeeService.cs
@@ -17,16 +17,25 @@ internal class DeleteEmployeeService
 
     public async Task RunAsync(EmployeeViewModel employee)
     {
-        AnsiConsole.WriteLine("Deleting employee....");
+        var confirmDelete = UserInput.GetConfirmation($"Are you sure you want to delete [green]{employee.FirstName} {employee.LastName}[/]?");
 
-        var result = await _employeeApiClient.DeleteAsync(new DeleteEmployeeCommand(employee.Id, employee.FirstName, employee.LastName));
+        Messages.PrintBlankLines(1);
+        if (confirmDelete)
+        {
+            var result = await _employeeApiClient.DeleteAsync(new DeleteEmployeeCommand(employee.Id, employee.FirstName, employee.LastName));
 
-        if (result.IsSuccess)
-            AnsiConsole.WriteLine($"Successfully deleted {employee.FirstName} {employee.LastName}");
+            if (result.IsSuccess)
+                AnsiConsole.WriteLine($"Successfully deleted {employee.FirstName} {employee.LastName}");
 
-        if (result.IsFailure)
-            Messages.OutputErrorMessage(result.Errors);
+            if (result.IsFailure)
+                Messages.OutputErrorMessage(result.Errors);
+        }
+        else
+        {
+            Messages.Cancelled($"Did not delete [green]{employee.FirstName} {employee.LastName}[/]");
+        }
 
+        Messages.PrintBlankLines(1);
         Messages.PressAnyKeyToContinue();
 
         return;
